+++ Assets/Scripts/Interface/IProcessable.cs
@@ -2,5 +2,5 @@
 using System.Collections;
 
 public interface IProcessable {
-	void SetProcess(float process, bool force = false);
+	void SetProcess(double process, bool force = false);
 }
